+++ tests/GridView/Column/CellTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\DataView\Tests\Column\Base;
+namespace Yiisoft\Yii\DataView\Tests\GridView\Column;
 
 use PHPUnit\Framework\TestCase;
 use Yiisoft\Yii\DataView\GridView\Column\Base\Cell;
